requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
decorator==5.0.9
Django==3.2.5
djangorestframework==3.12.4
fastcore==1.3.20
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
kora==0.9.19
matplotlib-inline==0.1.2
numpy==1.21.0
packaging==21.0
pandas==1.3.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webdriver-manager==3.4.2
